drivers: tz_driver: fix for pointer and integer type conversion errors when passing arguments to functions
../drivers/misc/mediatek/teei/300/tz_driver/backward_driver.c:203:48: warning: incompatible pointer to integer conversion passing 'void *' to parameter of type 'unsigned long' [-Wint-conversion]
        Invalidate_Dcache_By_Area((unsigned long)ptr, ptr + 4);
                                                      ^~~~~~~
../drivers/misc/mediatek/teei/300/tz_driver/include/teei_id.h:169:21: note: passing argument to parameter 'end' here
                                                unsigned long end)
                                                              ^
  LD      drivers/misc/mediatek/teei/300/tee/tee.o
../drivers/misc/mediatek/teei/300/tz_driver/irq_register.c:237:26: warning: unused variable 's' [-Wunused-variable]
        struct tz_driver_state *s = get_tz_drv_state();
                                ^
../drivers/misc/mediatek/teei/300/tz_driver/teei_client_main.c:731:22: warning: incompatible integer to pointer conversion passing 'unsigned long' to parameter of type 'const void *' [-Wint-conversion]
        memcpy(&versionlen, message_buff, sizeof(unsigned int));
                            ^~~~~~~~~~~~
../drivers/misc/mediatek/teei/300/tz_driver/teei_client_main.c:737:19: warning: incompatible integer to pointer conversion passing 'unsigned long' to parameter of type 'const void *' [-Wint-conversion]
                memcpy(version, message_buff + 4, versionlen);
                                ^~~~~~~~~~~~~~~~
../arch/arm64/include/asm/string.h:38:41: note: passing argument to parameter here
extern void *memcpy(void *, const void *, __kernel_size_t);
                                        ^

+++ drivers/misc/mediatek/teei/300/tz_driver/backward_driver.c
@@ -200,7 +200,7 @@ int __reetime_handle(struct service_handler *handler)
 	int time_type = 0;
 
 	ptr = handler->param_buf;
-	Invalidate_Dcache_By_Area((unsigned long)ptr, ptr + 4);
+	Invalidate_Dcache_By_Area((unsigned long)ptr, (unsigned long)(ptr + 4));
 	time_type = *((int *)ptr);
 	if (time_type == GET_UPTIME) {
 		get_monotonic_boottime(&tp);
